+++ tests/e2e/game/game-won-by-lovers.test.js
@@ -1,7 +1,6 @@
 const { describe, it, before, after } = require("mocha");
 const chai = require("chai");
 const chaiHttp = require("chai-http");
-const mongoose = require("mongoose");
 const app = require("../../../app");
 const Config = require("../../../config");
 const { resetDatabase } = require("../../../src/helpers/functions/Test");
@@ -50,17 +49,6 @@ describe("J - Tiny game of 4 players in which lovers win despite they're not on 
             .end((err, res) => {
                 expect(res).to.have.status(200);
                 game = res.body;
-                done();
-            });
-    });
-    it("🔐 Can't make a play if game's doesn't belong to user (POST /games/:id/play)", done => {
-        chai.request(app)
-            .post(`/games/${new mongoose.Types.ObjectId()}/play`)
-            .set({ Authorization: `Bearer ${token}` })
-            .send({ source: "all", action: "elect-sheriff" })
-            .end((err, res) => {
-                expect(res).to.have.status(401);
-                expect(res.body.type).to.equals("GAME_DOESNT_BELONG_TO_USER");
                 done();
             });
     });