+++ tests/e2e/game/index.js
@@ -11,4 +11,5 @@ describe("E2E - 🎲 Game tests", () => {
     require("./hunter-kills-mayor-on-day-time.test");
     require("./no-winner-game.test");
     require("./game-won-by-lovers.test");
+    require("./game-options.test");
 });